Server/requirements.txt
uvicorn~=0.25.0
PyYAML~=6.0.1
fastapi~=0.108.0
starlette~=0.32.0.post1
pydantic~=2.5.3
requests~=2.31.0